--- cypress/e2e/data-import/log-details/log-result-table-for-imported-multiple-items-in-multiple-holdings.cy.js
@@ -1,190 +1,0 @@
-import getRandomPostfix from '../../../support/utils/stringTools';
-import { DevTeams, TestTypes, Permissions, Parallelization } from '../../../support/dictionary';
-import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import SettingsMenu from '../../../support/fragments/settingsMenu';
-import {
-  FOLIO_RECORD_TYPE,
-  LOCATION_NAMES,
-  LOAN_TYPE_NAMES,
-  ITEM_STATUS_NAMES,
-  ACCEPTED_DATA_TYPE_NAMES,
-} from '../../../support/constants';
-import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
-import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
-import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
-import DataImport from '../../../support/fragments/data_import/dataImport';
-import TopMenu from '../../../support/fragments/topMenu';
-import Logs from '../../../support/fragments/data_import/logs/logs';
-import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
-import Users from '../../../support/fragments/users/users';
-import InstanceRecordView from '../../../support/fragments/inventory/instanceRecordView';
-import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
-import FieldMappingProfileView from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfileView';
-
-describe('data-import', () => {
-  describe('Log details', () => {
-    let user;
-    let instanceHRID;
-    const arrayOfHoldingsStatuses = ['Created (KU/CC/DI/M)', 'Created (KU/CC/DI/A)', 'Created (E)'];
-    const holdingsData = [
-      { permanentLocation: LOCATION_NAMES.MAIN_LIBRARY_UI, itemsQuqntity: 3 },
-      { permanentLocation: LOCATION_NAMES.ANNEX_UI, itemsQuqntity: 2 },
-      { permanentLocation: LOCATION_NAMES.ONLINE_UI, itemsQuqntity: 1 },
-    ];
-    const quantityOfCreatedItems = 6;
-    const quantityOfCreatedHoldings = 3;
-    const filePathForUpload = 'marcBibFileForC388505.mrc';
-    const marcFileName = `C388505 autotestFileName.${getRandomPostfix()}`;
-    const collectionOfMappingAndActionProfiles = [
-      {
-        mappingProfile: {
-          typeValue: FOLIO_RECORD_TYPE.HOLDINGS,
-          name: `C388505 Test multiple holdings.${getRandomPostfix()}`,
-          permanentLocation: '945$h',
-        },
-        actionProfile: {
-          typeValue: FOLIO_RECORD_TYPE.HOLDINGS,
-          name: `C388505 Test multiple holdings.${getRandomPostfix()}`,
-        },
-      },
-      {
-        mappingProfile: {
-          typeValue: FOLIO_RECORD_TYPE.ITEM,
-          name: `C388505 Test multiple items.${getRandomPostfix()}`,
-          materialType: '945$a',
-          permanentLoanType: LOAN_TYPE_NAMES.CAN_CIRCULATE,
-          status: ITEM_STATUS_NAMES.AVAILABLE,
-        },
-        actionProfile: {
-          typeValue: FOLIO_RECORD_TYPE.ITEM,
-          name: `C388505 Test multiple items.${getRandomPostfix()}`,
-        },
-      },
-    ];
-    const jobProfile = {
-      ...NewJobProfile.defaultJobProfile,
-      profileName: `C388505 Test multiple items.${getRandomPostfix()}`,
-      acceptedType: ACCEPTED_DATA_TYPE_NAMES.MARC,
-    };
-
-    before('login', () => {
-      cy.createTempUser([
-        Permissions.settingsDataImportEnabled.gui,
-        Permissions.moduleDataImportEnabled.gui,
-        Permissions.inventoryAll.gui,
-      ]).then((userProperties) => {
-        user = userProperties;
-
-        cy.login(userProperties.username, userProperties.password, {
-          path: SettingsMenu.mappingProfilePath,
-          waiter: FieldMappingProfiles.waitLoading,
-        });
-      });
-    });
-
-    after('delete test data', () => {
-      Users.deleteViaApi(user.userId);
-      JobProfiles.deleteJobProfile(jobProfile.profileName);
-      collectionOfMappingAndActionProfiles.forEach((profile) => {
-        ActionProfiles.deleteActionProfile(profile.actionProfile.name);
-        FieldMappingProfileView.deleteViaApi(profile.mappingProfile.name);
-      });
-      cy.getInstance({ limit: 1, expandAll: true, query: `"hrid"=="${instanceHRID}"` }).then(
-        (instance) => {
-          instance.items.forEach((item) => cy.deleteItemViaApi(item.id));
-          instance.holdings.forEach((holding) => cy.deleteHoldingRecordViaApi(holding.id));
-          InventoryInstance.deleteInstanceViaApi(instance.id);
-        },
-      );
-    });
-
-    it(
-      'C388505 Check the log result table for imported multiple items in multiple holdings (folijet)',
-      { tags: [TestTypes.smoke, DevTeams.folijet, Parallelization.nonParallel] },
-      () => {
-        // create mapping profiles
-        FieldMappingProfiles.openNewMappingProfileForm();
-        NewFieldMappingProfile.fillSummaryInMappingProfile(
-          collectionOfMappingAndActionProfiles[0].mappingProfile,
-        );
-        NewFieldMappingProfile.fillPermanentLocation(
-          collectionOfMappingAndActionProfiles[0].mappingProfile.permanentLocation,
-        );
-        NewFieldMappingProfile.save();
-        FieldMappingProfileView.closeViewMode(
-          collectionOfMappingAndActionProfiles[0].mappingProfile.name,
-        );
-
-        FieldMappingProfiles.openNewMappingProfileForm();
-        NewFieldMappingProfile.fillSummaryInMappingProfile(
-          collectionOfMappingAndActionProfiles[1].mappingProfile,
-        );
-        NewFieldMappingProfile.fillMaterialType(
-          collectionOfMappingAndActionProfiles[1].mappingProfile.materialType,
-        );
-        NewFieldMappingProfile.fillPermanentLoanType(
-          collectionOfMappingAndActionProfiles[1].mappingProfile.permanentLoanType,
-        );
-        NewFieldMappingProfile.fillStatus(
-          collectionOfMappingAndActionProfiles[1].mappingProfile.status,
-        );
-        NewFieldMappingProfile.save();
-        FieldMappingProfileView.closeViewMode(
-          collectionOfMappingAndActionProfiles[1].mappingProfile.name,
-        );
-
-        // create action profiles
-        collectionOfMappingAndActionProfiles.forEach((profile) => {
-          cy.visit(SettingsMenu.actionProfilePath);
-          ActionProfiles.create(profile.actionProfile, profile.mappingProfile.name);
-          ActionProfiles.checkActionProfilePresented(profile.actionProfile.name);
-        });
-
-        // create job profile
-        cy.visit(SettingsMenu.jobProfilePath);
-        JobProfiles.createJobProfile(jobProfile);
-        NewJobProfile.linkActionProfileByName('Default - Create instance');
-        NewJobProfile.linkActionProfile(collectionOfMappingAndActionProfiles[0].actionProfile);
-        NewJobProfile.linkActionProfile(collectionOfMappingAndActionProfiles[1].actionProfile);
-        NewJobProfile.saveAndClose();
-        JobProfiles.checkJobProfilePresented(jobProfile.profileName);
-
-        // upload .mrc file
-        cy.visit(TopMenu.dataImportPath);
-        // TODO delete function after fix https://issues.folio.org/browse/MODDATAIMP-691
-        DataImport.verifyUploadState();
-        DataImport.uploadFile(filePathForUpload, marcFileName);
-        JobProfiles.search(jobProfile.profileName);
-        JobProfiles.runImportFile();
-        JobProfiles.waitFileIsImported(marcFileName);
-        Logs.checkStatusOfJobProfile();
-        Logs.openFileDetails(marcFileName);
-        [
-          FileDetails.columnNameInResultList.srsMarc,
-          FileDetails.columnNameInResultList.instance,
-        ].forEach((columnName) => {
-          FileDetails.checkStatusInColumn(FileDetails.status.created, columnName);
-        });
-        FileDetails.verifyMultipleHoldingsStatus(
-          arrayOfHoldingsStatuses,
-          quantityOfCreatedHoldings,
-        );
-        FileDetails.verifyMultipleItemsStatus(quantityOfCreatedItems);
-
-        FileDetails.openInstanceInInventory('Created');
-        InventoryInstance.getAssignedHRID().then((initialInstanceHrId) => {
-          instanceHRID = initialInstanceHrId;
-        });
-        holdingsData.forEach((holdings) => {
-          InventoryInstance.checkIsHoldingsCreated([`${holdings.permanentLocation} >`]);
-          InventoryInstance.openHoldingsAccordion(`${holdings.permanentLocation} >`);
-          InstanceRecordView.verifyQuantityOfItemsRelatedtoHoldings(
-            holdings.permanentLocation,
-            holdings.itemsQuqntity,
-          );
-        });
-      },
-    );
-  });
-});
